Fix bug[#1514]: rtc test case fail when to set RTC_SWCNT_ON
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@3638 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/drivers/char/bfin_rtc.c
@@ -497,8 +497,8 @@ static int rtc_ioctl(struct inode *inode, struct file *file, unsigned int cmd,
 
 	case RTC_SWCNT_SET:	/* Read the periodic IRQ rate.  */
 		{
-			copy_from_user(&swcnttm, (unsigned long *)&arg,
-				       sizeof(unsigned long));
+			copy_from_user(&swcnttm, (unsigned short *)&arg,
+				       sizeof(unsigned short));
 
 			spin_lock_irq(&rtc_lock);
 			bfin_write_RTC_SWCNT(swcnttm);
